backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==25.1.0
sqlalchemy==2.0.45
psycopg2-binary==2.9.11
python-dotenv==1.0.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
requests==2.32.5
pydantic==2.10.5
pydantic-settings==2.7.1
pycryptodome==3.21.0
